modularize express and apollo
