+++ backend/src/utils/expressApp.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const isAuth = require("../middleware/jwt/is-auth");
+
+const app = express();
+app.use(isAuth);
+
+module.exports = app
